% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.Bayes.r
\name{EM.Bayes}
\alias{EM.Bayes}
\title{Estimation of the parameters of a mixture of two univariate normal distributions}
\usage{
EM.Bayes(x, Beta.pars = c(1, 1), lambda = 0, eps = 1e-08, verbose = TRUE)
}
\arguments{
\item{x}{a vector of data}

\item{Beta.pars}{parameters of the prior for pi}

\item{lambda}{parameter for the prior for sigma2^2/sigma1^2}

\item{eps}{criterion of convergence}

\item{verbose}{verbosity of the function}
}
\value{
a vector with named components
}
\description{
Estimation of the parameters of a mixture of two univariate normal distributions
}
\details{
This functions gives a Maximum A Posteriori for the parameter of a 
Gaussian mixture \eqn{pi*N(mu1, sigma1^2) + (1-pi)*N(mu2, sigma2^2)}.

The prior on pi is a Beta distribution. There's a prior on 
\eqn{r = sigma1^2/sigma2^2}, which is a \eqn{gamma(lambda + 1, lambda)}.

The default values for the prior parameters correspond to flat priors.
}
\examples{
set.seed(1);
x <- c( rnorm(100, 0), rnorm(100, 1) )
EM.Bayes(x)
# Recommanded non flat priors:
EM.Bayes(x, c(2,2), 1)
}
